Shams-backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dj-rest-auth==5.0.2
django-allauth==0.61.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
celery==5.3.6
redis==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
django-storages==1.14.2
boto3==1.34.34
openai==1.12.0
gunicorn==21.2.0
whitenoise==6.6.0
django-redis==5.4.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
Pillow==10.2.0 